feat: Add GPS data parser.
For now, only a GGL and RMC parser are available, but the code is
structured in a way that more can be added later. It also only retrieves
location data, nothing else. Date and time are not parsed.

The code needs a bit of refactoring, which will be done later.