feat: Remove `SimulatorDevice` as `accept` arg.
This is not needed anymore. We create a `std::unordered_map` in the
`ImGuiPresenter` class and pass that as a reference to the device
visitor. That way, we can keep "showing of a MCU" within the correct
class. This also meant I could _finally_ remove the `show` parameter
from `SimulatorDevice`, since that was no longer required there.

That finishes the implementation for a Simulator right now. A lot of
work has till to be done on the Simulator, but all taht was required now
is in there. Other features will come in later.

Fixes #69